Renamed courses-page to courses-overview, changed header
