+++ webpack/paths.js
@@ -33,7 +33,7 @@ const entry = {
 	"wp-seo-filter-explanation": "./wp-seo-filter-explanation.js",
 	"wp-seo-quick-edit-handler": "./wp-seo-quick-edit-handler.js",
 	"wp-seo-network-admin": "./wp-seo-network-admin.js",
-	"wp-seo-courses-overview": "./courses-page.js",
+	"wp-seo-courses-overview": "./courses-overview.js",
 };
 
 /**
